Re-add icons for related projects that have moved
This re-adds icons and links to bootc, podman desktop,
and youki.  All projects that have lived here in the past.

I have used them in the past, and would like to still be
able to get tos them from this page.

This also restores the deleted images from the logos
directory that should not have been removed in https://github.com/containers/.github/pull/18.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>